LayoutTests/fast/events/ios/autocorrect-with-apostrophe-expected.txt


This test verifies that "I'" can be autocorrected to "I’m". To manually run the test, focus the text area, place the caret at the end of the apostrophe, and select the "I’m" text candidate above the keyboard. Check that the text candidate is inserted as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS Observed input event of inputType insertReplacementText
PASS document.activeElement is editor
PASS editor.value is "I’m"
PASS successfullyParsed is true

TEST COMPLETE
